% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_management_step.R
\name{poststep2}
\alias{poststep2}
\title{poststep2 - Control chart for post management (Xbar-R)}
\usage{
poststep2(x)
}
\arguments{
\item{x}{: Xbar-R control chart data (subgroub size=4~5)}
}
\value{
: Xbar-R control chart
}
\description{
poststep2 - Control chart for post management (Xbar-R)
}
\examples{
 poststep2(pss6)
}
