% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prestep.R
\name{prestep1}
\alias{prestep1}
\title{prestep1 - Check the pre-treatment level}
\usage{
prestep1(x, y)
}
\arguments{
\item{x}{dataset}

\item{y}{y_values of dataset}
}
\value{
Data structure, Number of missing values, Outliers of data
}
\description{
prestep1 - Check the pre-treatment level
}
\examples{
prestep1(pss1,y)
}
