% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prestep.R
\name{prestep3}
\alias{prestep3}
\title{prestep3 - Dividing the dataset into test and verification sets}
\usage{
prestep3(x)
}
\arguments{
\item{x}{dataset}
}
\value{
trainset & testset
}
\description{
prestep3 - Dividing the dataset into test and verification sets
}
\examples{
prestep3(pss3)
}
