% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_step.R
\name{step1}
\alias{step1}
\title{step1 - Process capability analysis using data}
\usage{
step1(xST, xLT = NA, LSL = NA, USL = NA, Target = NA)
}
\arguments{
\item{xST}{short term data}

\item{xLT}{long term data (optional)}

\item{LSL}{lower spec limit}

\item{USL}{upper spec limit}

\item{Target}{target value}
}
\value{
process capability analysis
}
\description{
step1 - Process capability analysis using data
}
\examples{
short term : step1(xST=rnorm(30,100,2)  , LSL =95 , USL =105 , Target =100)
short & long term : step1(xST=rnorm(30,100,2)  ,xLT=rnorm(30,100,2), LSL =95 , USL =105 , Target =100)
}
