% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_step.R
\name{step2}
\alias{step2}
\title{step2 - Data exploration, Optimal model verification and confirmation}
\usage{
step2(x_train, x_test, y1, y2, y3)
}
\arguments{
\item{x_train}{: trainset}

\item{x_test}{: testset}

\item{y1}{: y expression of training data( y~. )}

\item{y2}{: y expression of test data( y~. )}

\item{y3}{: y value of test data( y )}
}
\value{
Linear& random forest model,Residual analysis result, Performance of both models
}
\description{
step2 - Data exploration, Optimal model verification and confirmation
}
\examples{
step2(pss3.1  , pss3.2 , pss3.2$throw_Distance~. , pss3.2$throw_Distance~. ,pss3.2$throw_Distance)
If there is no test set, put the training set in the test set field
}
