% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_step.R
\name{step3}
\alias{step3}
\title{step3 - Finding optimal conditions}
\usage{
step3(x)
}
\arguments{
\item{x:}{dataset}
}
\value{
Range, mean of dataset feature, Empty data frame
}
\description{
step3 - Finding optimal conditions
}
\examples{
step3(pss3)
}
