% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_step.R
\name{step4}
\alias{step4}
\title{step4 - Finding optimal results using  optimal conditions}
\usage{
step4(x_train, x_test, y1, y2, y3)
}
\arguments{
\item{x_train}{: trainset}

\item{x_test}{: newX(optimail conditon)}

\item{y1}{: y expression of training data( y~. )}

\item{y2}{: y expression of newX( y~. )}

\item{y3}{: y value of NewX( y )}
}
\value{

}
\description{
step4 - Finding optimal results using  optimal conditions
}
\examples{
 optimal results
step4(pss3.1  , pss3.2[1,] , pss3.2$throw_Distance~. , pss3.2$throw_Distance~. ,pss3.2$throw_Distance[1,])
If there is no test set, put the training set in the test set field
Enter the news example name and value in data.frame parentheses
newx=data.frame(throw_Distance=128, Rope_elasticity=3,Pulling_level=49, catapult_weight=46, catapult_Height=38, Throw_angle=48)
}
